vfs: Unhash and evict unused children dentries after rmdir
Patchwork-id: 121990
O-Subject: [RHEL6.8 fs PATCH] fs: Unhash and evict unused children dentries after rmdir
Bugzilla: 1241030
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Ian Kent <ikent@redhat.com>

BZ 1241030
https://bugzilla.redhat.com/show_bug.cgi?id=1241030

BREW 9488354
https://brewweb.devel.redhat.com/taskinfo?taskID=9488354

This commit does not have an upstream counterpart. Upstream code changed
significantly, so this is rhel6 specific problem.

Currently it's possibly with the following course of action to leave the
inode unable to be freed indefinitely. It has the consequence that it
might end up being freed by prune_cache (drop_caches, kswapd ...) which
can lead to a deadlock with file system freeze code, because if the file
system is frozen the iput_final will block on frozen file system while
holding s_umount which we need to take to thaw the file system.

Fix this by unhashing and evicting all unused dentries just after
successful directory remove to make sure that we can safely kill the
dentry.

Huge thanks to Al Viro for describing problem and proposing a
solution.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>