vfs: Prevent syncing frozen file system
Patchwork-id: 121992
O-Subject: [RHEL6.8 fs PATCH] fs: Prevent syncing frozen file system
Bugzilla: 1241791
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Dave Wysochanski <dwysocha@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 1241791
https://bugzilla.redhat.com/show_bug.cgi?id=1241791

This patch does not have upstream counterpart. This issue is present
upstream ATM, however patches that should fix the issue are waiting to
be merged (see the discussion
http://www.spinics.net/lists/linux-fsdevel/msg87735.html).
Unfortunately that upstream fix is part of bigger rewrite which we do
not want to backport for RHEL just yet. Instead we have this much
simpler solution that could be replaced later.

Currently we can end up in a deadlock because of broken
sb_start_write -> s_umount ordering.

The race goes like this:

 - write the file
 - unlink the file - final_iput will not be calles as file is opened
 - freeze the file system
 - Now simultaneously close the file and call sync (or syncfs on that
   particular file system). Sync will get to wait_sb_inodes() where it will
   grab the referece to the inode (__iget()) and later to call iput().
   If we manage to close the file and drop the reference in between those
   calls sync will attempt to do a iput_final() because the inode is now
   unlinked and we're holding the last reference to it. This will
   however block on a frozen file system (ext4_delete_inode for
   example).

Note that I've not been able to reproduce the issue, I've only seen this
happen once. However with some instrumentation (like msleep() in the
wait_sb_inodes() it can be achieved.

Fix this by skipping the sync (or syncfs) on frozen file system as it
really should be clean anyway.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Kurt Stutsman <kstutsma@redhat.com>