tracing: Simplify test for function_graph tracing start point
Patchwork-id: 122173
O-Subject: [PATCH RHEL6 BZ1063695 1/3] tracing: Simplify test for function_graph tracing start point
Bugzilla: 1063695
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Pratyush Anand <panand@redhat.com>

Bugzilla: 1063695
https://bugzilla.redhat.com/show_bug.cgi?id=1063695

upstream
========
commit ea2c68a08fedb5053ba312d661e47df9f4d72411
Author: Lai Jiangshan <laijs@cn.fujitsu.com>
Date:   Wed Jan 13 19:38:30 2010 +0800

description
===========
In the function graph tracer, a calling function is to be traced
only when it is enabled through the set_graph_function file,
or when it is nested in an enabled function.

Current code uses TSK_TRACE_FL_GRAPH to test whether it is nested
or not. Looking at the code, we can get this:
(trace->depth > 0) <==> (TSK_TRACE_FL_GRAPH is set)

trace->depth is more explicit to tell that it is nested.
So we use trace->depth directly and simplify the code.

No functionality is changed.
TSK_TRACE_FL_GRAPH is not removed yet, it is left for future usage.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>