perf hists browser: Override ui_browser refresh_dimensions method
Patchwork-id: 123561
O-Subject: [PATCH RHEL6 BZ1216217 054/341] perf hists browser: Override ui_browser refresh_dimensions method
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit 357cfff1c2173b7aedc0e521a2146b78396171e0
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jul 1 17:01:01 2014 -0300

description
===========
This requires some more work so that we can really just use the width of
current entries when we want to partition the screen.

Right now its just a prep patch so that we can have where to update
ui_browser->rows when introducing the column headers line, that will be
togglable, so we need to update it everytime we refresh the dimensions
of the browser.