perf session: Flag if the event stream is entirely in memory
Patchwork-id: 123634
O-Subject: [PATCH RHEL6 BZ1216217 088/341] perf session: Flag if the event stream is entirely in memory
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit 919d86d3a3109d7d4f0d7347f34711ee2f8e6609
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Jul 14 13:02:51 2014 +0300

description
===========
Flag if the event stream is a file that has been mmapped in one go.

This is useful, for example, if a tool needs to keep an event for later
reference.  If the new flag is set, a pointer to the event can be
retained, otherwise the event must be copied.