perf tools: Add dsos__hit_all()
Patchwork-id: 123642
O-Subject: [PATCH RHEL6 BZ1216217 110/341] perf tools: Add dsos__hit_all()
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit 1f625b0b3dd641b74ba21640a1fea19a3dd893a2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Jul 22 16:17:23 2014 +0300

description
===========
Add ability to mark all dsos as hit.

This is needed in the case of Instruction Tracing.  It takes so long to
decode an Instruction Trace that it is not worth doing just to determine
which dsos are hit.  A later patch takes this into use.