perf tools: Flush ordered events in case of allocation failure
Patchwork-id: 123668
O-Subject: [PATCH RHEL6 BZ1216217 144/341] perf tools: Flush ordered events in case of allocation failure
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit d40b4a15ab2bfcfa7d946b69ca1f12c93b22d169
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Aug 1 13:01:04 2014 -0300

description
===========
In previous patches we added a limit for ordered events queue allocation
size. If we reach this size we need to flush (part of) the queue to get
some free buffers.

The current functionality is not affected, because the limit is hard
coded to (u64) -1. The configuration code for size will come in
following patches.