perf tools: Fix column alignment when headers aren't shown on TUI
Patchwork-id: 123677
O-Subject: [PATCH RHEL6 BZ1216217 162/341] perf tools: Fix column alignment when headers aren't shown on TUI
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit 59dc9f2534569d11a55c8b5dbe93c36f2b2fa506
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Jul 31 14:47:41 2014 +0900

description
===========
If user sets ui.show-headers config option to false, it didn't calculate
default column width so it broke the alignment.  This is because it does
the calculation just before showing headers.

Move it to the beginning of the hist browser so that it can be called
regardless of the config option.