perf evlist: Add 'system_wide' option
Patchwork-id: 123715
O-Subject: [PATCH RHEL6 BZ1216217 197/341] perf evlist: Add 'system_wide' option
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit bf8e8f4b832972c76d64ab2e2837a48397144887
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Jul 31 09:00:51 2014 +0300

description
===========
Add an option to cause a selected event to be opened always without a
pid when configured by perf_evsel__config().

This is needed when using the sched_switch tracepoint to follow object
code execution.

sched_switch occurs before the task switch and so it cannot record it in
a context limited to that task.  Note that also means that sched_switch
is useless when capturing data per-thread, as is the 'context-switches'
software event for the same reason.