perf evlist: We need to poll all event file descriptors
Patchwork-id: 123769
O-Subject: [PATCH RHEL6 BZ1216217 260/341] perf evlist: We need to poll all event file descriptors
Bugzilla: 1216217
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1216217
https://bugzilla.redhat.com/show_bug.cgi?id=1216217

upstream
========
commit 033fa713db66b96d5779e6a93d50ff821bc1abd2
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Mon Sep 8 12:55:12 2014 -0300

description
===========
Because we want to notice when they get POLLHUP'ed, so that we can
figure out when all threads exited in a workload being monitored.

We can't just monitor the fds that were mmaped, we need to notice when
all the fds that were PERF_EVENT_IOC_SET_OUTPUT'ed too, because the mmap
stays even after the fd that originally was used to do the mmap call
went away, its only when all the set-output fds for a mmap are gone that
the mmap is.