Add ability to add extra text to description
Patchwork-id: 124971
O-Subject: [RHEL6 RFC] Add ability to add extra text to description.
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

In SEG when we build custom kernels to customer to test fixes we tend to
add disclaimer to the description of the built rpm to clearly say that
this rpm is not supported and is provided for testing purposes only. The
usual way of doing this is to create branch in dist-git, manually modify
the spec file, add patch(es) and brew it.

I personally like the make rh-brew(-scratch) approach which is much
faster and I don't have to leave my local git checkout. I still have to
manually modify the specfile template so that the disclaimer is present
but I forget from times to times.

Because of this and because I like to automate manual stuff I would like
to propose below addition. It's simply modification of spec template and
create-patches.sh that look for 'localdesc' file in root (similar to
localversion already present) and adds content of this file to spec
file.

Any suggestions/concerns are welcome.

Below patch tested on
https://brewweb.devel.redhat.com/taskinfo?taskID=9821367

Unfortunately the built happened some time ago and is no longer
available for download but srpm can be found here
http://download.devel.redhat.com/brewroot/scratch/mvyletel/task_9821367/

Signed-off-by: Milos Vyletel <milos@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>