Sanitize get_callchain_buffer()
Patchwork-id: 124997
O-Subject: [PATCH RHEL6.8 BZ1252428 02/28] perf: Sanitize get_callchain_buffer()
Bugzilla: 1252428
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1252428
https://bugzilla.redhat.com/show_bug.cgi?id=1252428

upstream
========
commit 90983b16078ab0fdc58f0dab3e8e3da79c9579a2
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Tue Jul 23 02:31:00 2013 +0200

description
===========
In case of allocation failure, get_callchain_buffer() keeps the
refcount incremented for the current event.

As a result, when get_callchain_buffers() returns an error,
we must cleanup what it did by cancelling its last refcount
with a call to put_callchain_buffers().

This is a hack in order to be able to call free_event()
after that failure.

The original purpose of that was to simplify the failure
path. But this error handling is actually counter intuitive,
ugly and not very easy to follow because one expect to
see the resources used to perform a service to be cleaned
by the callee if case of failure, not by the caller.

So lets clean this up by cancelling the refcount from
get_callchain_buffer() in case of failure. And correctly free
the event accordingly in perf_event_alloc().

Signed-off-by: Aristeu Rozanski <aris@redhat.com>