Add ->count function to read per-package counters
Patchwork-id: 125002
O-Subject: [PATCH RHEL6.8 BZ1252428 07/28] perf: Add ->count() function to read per-package counters
Bugzilla: 1252428
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1252428
https://bugzilla.redhat.com/show_bug.cgi?id=1252428

upstream
========
commit eacd3ecc34472ce3751eedfc94e44c7cc6eb6305
Author: Matt Fleming <matt.fleming@intel.com>
Date: Fri Jan 23 18:45:41 2015 +0000

description
===========
For PMU drivers that record per-package counters, the ->count variable
cannot be used to record an accurate aggregated value, since it's not
possible to perform SMP cross-calls to cpus on other packages from the
context in which we update ->count.

Introduce a new optional ->count() accessor function that can be used to
customize how values are collected. If a PMU driver doesn't provide a
->count() function, we fallback to the existing code.

There is necessarily a window of staleness with this approach because
the task that generated the counter value may not have been scheduled by
the cpu recently.

An alternative and more complex approach would be to use a hrtimer to
periodically refresh the values from a more permissive scheduling
context. So, we're trading off complexity for accuracy.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>