x86/intel: Implement LRU monitoring ID allocation for CQM
Patchwork-id: 125009
O-Subject: [PATCH RHEL6.8 BZ1252428 14/28] perf/x86/intel: Implement LRU monitoring ID allocation for CQM
Bugzilla: 1252428
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1252428
https://bugzilla.redhat.com/show_bug.cgi?id=1252428

upstream
========
commit 35298e554c74b7849875e3676ba8eaf833c7b917
Author: Matt Fleming <matt.fleming@intel.com>
Date: Fri Jan 23 18:45:45 2015 +0000

description
===========
It's possible to run into issues with re-using unused monitoring IDs
because there may be stale cachelines associated with that ID from a
previous allocation. This can cause the LLC occupancy values to be
inaccurate.

To attempt to mitigate this problem we place the IDs on a least recently
used list, essentially a FIFO. The basic idea is that the longer the
time period between ID re-use the lower the probability that stale
cachelines exist in the cache.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>