x86/intel: Support task events with Intel CQM
Patchwork-id: 125010
O-Subject: [PATCH RHEL6.8 BZ1252428 15/28] perf/x86/intel: Support task events with Intel CQM
Bugzilla: 1252428
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1252428
https://bugzilla.redhat.com/show_bug.cgi?id=1252428

upstream
========
commit bfe1fcd2688f557a6b6a88f59ea7619228728bd7
Author: Matt Fleming <matt.fleming@intel.com>
Date: Fri Jan 23 18:45:46 2015 +0000

RHEL6
=====
- using local version of cgroup_is_descendant backported from
  78574cf981cd cgroup: implement cgroup_is_descendant()
  RHEL6 still exports old version checking the cgroup and task.
- moving uapi changes into include/linux/perf_event.h
- small conflict in core.c due to missing breakpoint support in RHEL6

description
===========
Add support for task events as well as system-wide events. This change
has a big impact on the way that we gather LLC occupancy values in
intel_cqm_event_read().

Currently, for system-wide (per-cpu) events we defer processing to
userspace which knows how to discard all but one cpu result per package.

Things aren't so simple for task events because we need to do the value
aggregation ourselves. To do this, we defer updating the LLC occupancy
value in event->count from intel_cqm_event_read() and do an SMP
cross-call to read values for all packages in intel_cqm_event_count().
We need to ensure that we only do this for one task event per cache
group, otherwise we'll report duplicate values.

If we're a system-wide event we want to fallback to the default
perf_event_count() implementation. Refactor this into a common function
so that we don't duplicate the code.

Also, introduce PERF_TYPE_INTEL_CQM, since we need a way to track an
event's task (if the event isn't per-cpu) inside of the Intel CQM PMU
driver.  This task information is only availble in the upper layers of
the perf infrastructure.

Other perf backends stash the target task in event->hw.*target so we
need to do something similar. The task is used to determine whether
events should share a cache group and an RMID.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>