x86/intel: Perform rotation on Intel CQM RMIDs
Patchwork-id: 125011
O-Subject: [PATCH RHEL6.8 BZ1252428 16/28] perf/x86/intel: Perform rotation on Intel CQM RMIDs
Bugzilla: 1252428
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1252428
https://bugzilla.redhat.com/show_bug.cgi?id=1252428

upstream
========
commit bff671dba7981195a644a5dc210d65de8ae2d251
Author: Matt Fleming <matt.fleming@intel.com>
Date: Fri Jan 23 18:45:47 2015 +0000

RHEL6
=====
- switched raw_spinlock_t into spinlock_t plus related calls

description
===========
There are many use cases where people will want to monitor more tasks
than there exist RMIDs in the hardware, meaning that we have to perform
some kind of multiplexing.

We do this by "rotating" the RMIDs in a workqueue, and assigning an RMID
to a waiting event when the RMID becomes unused.

This scheme reserves one RMID at all times for rotation. When we need to
schedule a new event we give it the reserved RMID, pick a victim event
from the front of the global CQM list and wait for the victim's RMID to
drop to zero occupancy, before it becomes the new reserved RMID.

We put the victim's RMID onto the limbo list, where it resides for a
"minimum queue time", which is intended to save ourselves an expensive
smp IPI when the RMID is unlikely to have a occupancy value below
__intel_cqm_threshold.

If we fail to recycle an RMID, even after waiting the minimum queue time
then we need to increment __intel_cqm_threshold. There is an upper bound
on this threshold, __intel_cqm_max_threshold, which is programmable from
userland as /sys/devices/intel_cqm/max_recycling_threshold.

The comments above __intel_cqm_rmid_rotate() have more details.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>