netpoll: Close race condition between poll_one_napi and napi_disable
Patchwork-id: 125040
O-Subject: [RHEL6.8PATCH] netpoll: Close race condition between poll_one_napi and napi_disable
Bugzilla: 1252212
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Paolo Abeni <pabeni@redhat.com>

Drivers might call napi_disable while not holding the napi instance poll_lock.
In those instances, its possible for a race condition to exist between
poll_one_napi and napi_disable.  That is to say, poll_one_napi only tests the
NAPI_STATE_SCHED bit to see if there is work to do during a poll, and as such
the following may happen:

CPU0				CPU1
ndo_tx_timeout			napi_poll_dev
 napi_disable			 poll_one_napi
  test_and_set_bit (ret 0)
				  test_bit (ret 1)
   reset adapter		   napi_poll_routine

If the adapter gets a tx timeout without a napi instance scheduled, its possible
for the adapter to think it has exclusive access to the hardware  (as the napi
instance is now scheduled via the napi_disable call), while the netpoll code
thinks there is simply work to do.  The result is parallel hardware access
leading to corrupt data structures in the driver, and a crash.

Additionaly, there is another, more critical race between netpoll and
napi_disable.  The disabled napi state is actually identical to the scheduled
state for a given napi instance.  The implication being that, if a napi instance
is disabled, a netconsole instance would see the napi state of the device as
having been scheduled, and poll it, likely while the driver was dong something
requiring exclusive access.  In the case above, its fairly clear that not having
the rings in a state ready to be polled will cause any number of crashes.

The fix should be pretty easy.  netpoll uses its own bit to indicate that that
the napi instance is in a state of being serviced by netpoll (NAPI_STATE_NPSVC).
We can just gate disabling on that bit as well as the sched bit.  That should
prevent netpoll from conducting a napi poll if we convert its set bit to a
test_and_set_bit operation to provide mutual exclusion

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
CC: "David S. Miller" <davem@davemloft.net>
CC: jmaxwell@redhat.com
Tested-by: jmaxwell@redhat.com
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 2d8bff12699abc3a9bf886bb0b79f44d94d81496)

Note: Upstreams patch is slightly different due to the fact that napi_disable
has been made a full function rather than an inline function. I also realized in
the process of backporting that the unlocks aren't in the same order as the
locks, so I corrected that, and will be doing so upstream shortly

BZ: 1252212
Testing: Tested by the reporter
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9882497

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>