random: statically compute poolbitshift, poolbytes, poolbits
Patchwork-id: 125654
O-Subject: [RHEL6 PATCH 1/3] random: statically compute poolbitshift, poolbytes, poolbits
Bugzilla: 1110915
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

RHEL6 Bugzilla #1110915

Upstream commit: 9ed17b70b409dc48c134a80b5a6df582ba759de2

Use a macro to statically compute poolbitshift (will be used in a
subsequent patch), poolbytes, and poolbits.  On virtually all
architectures the cost of a memory load with an offset is the same as
the one of a memory load.

It is still possible for this to generate worse code since the C
compiler doesn't know the fixed relationship between these fields, but
that is somewhat unlikely.

Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>