crypto: Build accelerated versions of SHA into the kernel
Patchwork-id: 125747
O-Subject: [RHEL6 PATCH v2] crypto: Build accelerated versions of SHA into the kernel
Bugzilla: 1229864
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

RHEL6 Bugzilla #1229864

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9969616

The crypto layer uses module aliases to ensure that all versions
of a given algorithm are considered.  This means that if you
want to build an algorithm into the kernel statically, then all
implementations of that algorithm that may be used need to be
built in as well to ensure that the best version is chosen.

We were building the generic SHA implementations statically but
not the accelerated versions of SHA, meaning that the accelerated
SHAs never got used.

This patch makes thes the accelerated SHAs built-in as well.

Signed-off-by: Herbert Xu <herbert.xu@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>