svcrpc: support multiple-fragment rpc's
Patchwork-id: 125818
O-Subject: [Patch RHEL6 fs 08/11] svcrpc: support multiple-fragment rpc's
Bugzilla: 987452
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

Upstream: 836fbadb96a8308e6283eee5c7b3bdae818b58ca
Bugzilla: 1250739

Over TCP, RPC's are preceded by a single 4-byte field telling you how
long the rpc is (in bytes).  The spec also allows you to send an RPC in
multiple such records (the high bit of the length field is used to tell
you whether this is the final record).

We've survived for years without supporting this because in practice the
clients we care about don't use it.  But the userland rpc libraries do,
and every now and then an experimental client will run into this.  (Most
recently I noticed it while trying to write a pynfs check.)  And we're
really on the wrong side of the spec here--let's fix this.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>