svcrpc: don't error out on small tcp fragment
Patchwork-id: 125823
O-Subject: [Patch RHEL6 fs 11/11] svcrpc: don't error out on small tcp fragment
Bugzilla: 987452
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

Upstream: 1f691b07c5dc51b2055834f58c0f351defd97f27
Bugzilla: 1250739

Though clients we care about mostly don't do this, it is possible for
rpc requests to be sent in multiple fragments.  Here we have a sanity
check to ensure that the final received rpc isn't too small--except that
the number we're actually checking is the length of just the final
fragment, not of the whole rpc.  So a perfectly legal rpc that's
unluckily fragmented could cause the server to close the connection
here.

Cc: stable@vger.kernel.org
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>