intel_pstate, convert kernel parameters into module parameters for RHEL6
Patchwork-id: 126024
O-Subject: [RHEL6.8 PATCH BZ 1246961 57/57] intel_pstate, convert kernel parameters into module parameters for RHEL6
Bugzilla: 1246961
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1246961

RHEL_only.

In the upstream kernel, intel_pstate is a built-in.  In RHEL6 (since the
default x86 cpu frequency driver is acpi-cpufreq) we have had to make it
a module so that it can be loaded during runtime.

The intel_pstate kernel parameters need to be converted into module
parameters so that the force, no_hwp, and, hwp_only options can be used
with the driver in RHEL6.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: Steve Best <sbest@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>