fs-cache: When submitting an op, cancel it if the target object is dying
Patchwork-id: 126160
O-Subject: [RHEL6 PATCH 26/36] FS-Cache: When submitting an op, cancel it if the target object is dying
Bugzilla: 1277622
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

When submitting an operation, prefer to cancel the operation immediately
rather than queuing it for later processing if the object is marked as dying
(ie. the object state machine has reached the KILL_OBJECT state).

Whilst we're at it, change the series of related test_bit() calls into a
READ_ONCE() and bitwise-AND operators to reduce the number of load
instructions (test_bit() has a volatile address).

[Note: For RHEL-7 calls to READ_ONCE() have been converted to ACCESS_ONCE()]

Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Steve Dickson <steved@redhat.com>
Acked-by: Jeff Layton <jeff.layton@primarydata.com>
Original-commit: 30ceec6284129662efc3a1e7675b2bd857a046fe
Signed-off-by: Aristeu Rozanski <aris@redhat.com>