fs-cache: Put an aborted initialised op so that it is accounted correctly
Patchwork-id: 126168
O-Subject: [RHEL6 PATCH 34/36] FS-Cache: Put an aborted initialised op so that it is accounted correctly
Bugzilla: 1277622
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Call fscache_put_operation() or a wrapper on any op that has gone through
fscache_operation_init() so that the accounting shown in /proc is done
correctly, specifically fscache_n_op_release.

fscache_put_operation() therefore now allows an op in the INITIALISED state as
well as in the CANCELLED and COMPLETE states.

Note that this means that an operation can get put that doesn't have its
->object pointer filled in, so anything that depends on the object needs to be
conditional in fscache_put_operation().

Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Steve Dickson <steved@redhat.com>
Acked-by: Jeff Layton <jeff.layton@primarydata.com>
Original-commit: a39caadf06879017cb9a8c5c5cb4fc4ccb213275
Signed-off-by: Aristeu Rozanski <aris@redhat.com>