gfs2: Protect freeing directory hash table with i_lock spin_lock
Patchwork-id: 126214
O-Subject: [RHEL6.8 fs PATCH] GFS2: Protect freeing directory hash table with i_lock spin_lock
Bugzilla: 1250663
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

Hi,

This patch is for bug #1250663. The problem is a double-free of GFS2
directory hash information in memory.

It was tested on system gfs-a16c-04.mpc.lab.eng.bos.redhat.com, but I was
not able to recreate the problem, so it was just a sanity test to make sure
the patch didn't break anything.

A kernel build was done here:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10068082

The upstream patch is here:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=c36b97e9430defab4b52ee370c18745343d92f6d

Patch text:

This patch changes function gfs2_dir_hash_inval so that it uses
the i_lock spin_lock to protect the in-core hash table, i_hash_cache.
This will prevent double-frees due to a race between gfs2_clear_inode
and inode invalidation.

rhbz#1250663

Signed-off-by: Bob Peterson <rpeterso@redhat.com>