Revert "Update USB default wakeup settings"
Patchwork-id: 126332
O-Subject: [RHEL6 PATCH] Revert "[usb] Update USB default wakeup settings"
Bugzilla: 1236589
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/1236589
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10072076
Tested: on a virt guest which showed problem.
        on a Dell box with the original problem, no regressions

This reverts commit 1ebd39678340f5de9eb1d543eb14ce6410441fba.

The above commit was 1 of 3 patches in an attempt by me to get the auto power
settings working for bz 698729.  Looking through the commit again, I was a
little to agressive and pulled in a host controller patch that changed the
default settings.  This behaviour changed caused the controller to constantly
send messages to the cpu wanting attention in a guest.

A little bit of history.  RHEL-6's auto-suspend framework is still immature.
Even worse, for RHEL-6 usb it was using its own framework.  USB auto-suspend
migrated to the kernel's framework post-RHEL-6.  As a result a lot of this
isn't going to work correctly.  I throw in random patches to help keep
customers happy.

Reverting this patch, retains the fix for the original problem and allows the
virtual cpu to go to idle correctly.  The original fix is resolved by the other
2 patches in the original patch series.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>