ftrace: Do not trace copy_user_generic
Patchwork-id: 126343
O-Subject: [PATCH RHEL 6 BZ 1278884] ftrace: Do not trace copy_user_generic()
Bugzilla: 1278884
RH-Acked-by: Clark Williams <williams@redhat.com>
RH-Acked-by: Luis Claudio R. Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Daniel Bristot de Oliveira <bristot@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1278884
Upstream Status: N/A (this code has been completely rewritten there)
Tested: Stumbled over this while testing another BZ, this fixed it there

When modifying the text to enable tracing of a function which requires
that
the nop stub in the preamble of the function gets converted to a call to the
function tracer, it must be done atomically and no other CPU should be executing
that code. If another CPU does, then it could crash due to the modification
crossing cache lines. Ftrace does this by calling stop_machine() to make sure
that only one CPU is active when the update is made. But this does not prevent
NMIs from other CPUs from running. To handle that case, the ftrace code does
a dance with the NMIs where if an update is being made, the first thing the
NMI will do in the NMI handler is to do the update itself, to make sure that the
code is coherent on its CPU.

The issue is that this modification is done with probe_kernel_write() which
uses copy_user_generic(), which can be traced by the function tracer. What
happens is that copy_user_generic() gets modified by one NMI, but then it is
used by another NMI on another CPU which triggers the bug because the
copy_user_generic() itself was modified. The the system crashes.

Easiest solution is to simply not trace copy_user_generic() and then it will
not be modified either.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>