md: clear Blocked flag on failed devices when array is read-only
Patchwork-id: 126582
O-Subject: [RHEL6.8 PATCH 1/1] md: clear Blocked flag on failed devices when array is read-only.
Bugzilla: 1177913
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Neil Brown <neilb@suse.de>

The Blocked flag indicates that a device has failed but that this
fact hasn't been recorded in the metadata yet.  Writes to such
devices cannot be allowed until the metadata has been updated.

On a read-only array, the Blocked flag will never be cleared.
This prevents the device being removed from the array.

If the metadata is being handled by the kernel
(i.e. !mddev->external), then we can be sure that if the array is
switch to writable, then a metadata update will happen and will
record the failure.  So we don't need the flag set.

If metadata is externally managed, it is upto the external manager
to clear the 'blocked' flag.

Reported-by: XiaoNi <xni@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit ab16bfc732c436658d13455f28b0b4a2608a7476)
Signed-off-by: Xiao Ni <xni@redhat.com>

Signed-off-by: Aristeu Rozanski <aris@redhat.com>