ext4: fix label jump on broken fs in ext4_fill_super
Patchwork-id: 126610
O-Subject: [RHEL6.8 fs PATCH] ext4: fix label jump on broken fs in ext4_fill_super
Bugzilla: 1266862
CVE: CVE-2015-8324
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Jiaying Zhang <jiayingz@google.com>

BZ 1266862
https://bugzilla.redhat.com/show_bug.cgi?id=1266862

BREW 10092291
https://brewweb.devel.redhat.com/taskinfo?taskID=10092291

This commit is just a very minor (and seemingly unrelated) part
of the upstream commit 744692dc059845b2a3022119871846e74d4f6e11 because
we're not intending to backport the functionality this commit provides
but rather just a fix of the wrong label jump in ext4_fill_super(). Also
not that even the fix itself is a bit different because the upstream
changed and we still need one more place to change the label. Also I am
going to rewrite the description since it has nothing to do with the
original commit.

	Currently if we encounter error in some places in ext4_fill_super() we
	might jump to a wrong label (failed_mount4) and try to destroy
	not-yet-created workqueue and hence end up with NULL pointer
	dereference. Fix it by changing the wrong label in some places to
	failed_mount_wq.

	Signed-off-by: Lukas Czerner <lczerner@redhat.com>

Signed-off-by: Aristeu Rozanski <aris@redhat.com>