xfs: add a sysfs kset
Patchwork-id: 126617
O-Subject: [RHEL6 PATCH v4 03/17] xfs: add a sysfs kset
Bugzilla: 1205640
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>

From: Brian Foster <bfoster@redhat.com>

commit 3d8712265c26546823b38eb97487262500ff13db
Author: Brian Foster <bfoster@redhat.com>
Date:   Tue Jul 15 07:41:37 2014 +1000

Create a sysfs kset to contain all sub-objects associated with the XFS
module. The kset is created and removed on module initialization and
removal respectively. The kset uses fs_obj as a parent. This leads to
the creation of a /sys/fs/xfs directory when the kset exists.

Signed-off-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>