xfs: add xlog sysfs kobject and attribute handlers
Patchwork-id: 126623
O-Subject: [RHEL6 PATCH v4 05/17] xfs: add xlog sysfs kobject and attribute handlers
Bugzilla: 1205640
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>

From: Brian Foster <bfoster@redhat.com>

commit baff4e44b92aad363dcce176b907a369ee8b348e
Author: Brian Foster <bfoster@redhat.com>
Date:   Tue Jul 15 08:07:29 2014 +1000

Embed a kobject into the xfs log data structure (xlog). This creates a
'log' subdirectory for every XFS mount instance in sysfs. The lifecycle
of the log kobject is tied to the lifecycle of the log.

Also define a set of generic attribute handlers associated with the log
kobject in preparation for the addition of attributes.

Signed-off-by: Brian Foster <bfoster@redhat.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>