nfsv3: Use the readdir fileid as the mounted-on-fileid
Patchwork-id: 126765
O-Subject: [RHEL6.8 PATCH fs BZ1223978 2/3] NFSv3: Use the readdir fileid as the mounted-on-fileid
Bugzilla: 1223978
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

When we call readdirplus, set the fileid normally returned by readdir
as the mounted-on-fileid, since that is commonly the case if there is
a mountpoint. To ensure that we get it right, we only set the flag if
the readdir fileid differs from the one returned in the readdirplus
attributes.

This again means that we can avoid the issues described in commit
2ef47eb1aee17 ("NFS: Fix use of nfs_attr_use_mounted_on_fileid()"),
which only fixed NFSv4.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1223978
Upstream Id: 1ae04b252351188cfa66af1b8b0628512a72dd7b
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>

Signed-off-by: Aristeu Rozanski <aris@redhat.com>