clean up the device handling for the bmc device
Patchwork-id: 127043
O-Subject: [RHEL6.8 PATCH 05/34] ipmi: clean up the device handling for the bmc device
Bugzilla: 1274302
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1274302
brew: https://brewweb.devel.redhat.com/taskinfo?taskID=10117391

NOTES:
struct bmc_device is not exposed to 3rd parties, but the RHEL6 kabi
checker complains about it, because a pointer to it is embedded in
struct ipmi_smi, which is also not exposed to 3rd parties. However,
because a pointer to ipmi_smi is exposed, the RHEL6 checker insists
on consistency of its signature and the signatures of all its fields,
even if they are just pointers. This overreaching characteristic of
the kabi checker was fixed in RHEL7.

Cherry-picked from the following upstream commit.

commit 16639eb08a6947bef8f4d78e22b6010ba4c3cb90
Author: Corey Minyard <cminyard@mvista.com>
Date:   Fri Oct 10 21:54:03 2014 -0500

    ipmi: clean up the device handling for the bmc device

    Embed the platform device in the bmc device instead of externally
    allocating it, use more proper form for creating the device
    attributes, and other general cleanups.

    Signed-off-by: Corey Minyard <cminyard@mvista.com>

Signed-off-by: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>