hpsa: add masked physical devices into struct h's dev array
Patchwork-id: 127222
O-Subject: [RHEL6.8 e-stor PATCH 02/58] hpsa: add masked physical devices into h->dev[] array
Bugzilla: 1247729
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

This upstream commit was manually applied.   It is pretty much the same as
is found upstream at:

https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=41ce4c355765a5e0a8e1c8ff8d7257160bbae93d

However, hpsa_slave_configure() has been modified to use
hpsa_adjust_queue_depth() for RHEL6.   Upstream commit used
scsi_change_queue_depth() which is not available in RHEL6.

The following is text from the upstream commit:

Cache the ioaccel handle so that when we need to abort commands sent
down the ioaccel2 path, we can look up the LUN ID in h->dev[] instead of
having to do I/O to the controller.

Add a field to elements in h->dev[] to keep track of how the device is exposed
to the SCSI mid layer: Not at all, without an upper level driver
(no_uld_attach) or normally exposed.

Since masked physical devices are now present in h->dev[] array
it would be perfectly possible to do

	echo scsi add-single-device 2 2 0 0 > /proc/scsi/scsi

and bring them online.  This was previously not allowed for masked
physical devices.

Ensure that the mapping of physical disks to logical drives gets updated in a
consistent way when a RAID migration occurs and is not touched until updates
to it are complete.

now instead of doing CISS_REPORT_PHYSICAL to get the LUNID for
the physical disk in hpsa_get_pdisk_of_ioaccel2(), just get
it out of h->dev[] where we already have it cached.

do not touch phys_disk[] for ioaccel enabled logical drives during rescan

Reviewed-by: Scott Teel <scott.teel@pmcs.com>
Reviewed-by: Kevin Barnett <kevin.barnett@pmcs.com>
Reviewed-by: Tomas Henzl <thenzl@redhat.com>
Reviewed-by: Hannes Reinecke <hare@Suse.de>
Signed-off-by: Don Brace <don.brace@pmcs.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: James Bottomley <JBottomley@Odin.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>