hpsa: move lockup_detected attribute to host attr
Patchwork-id: 127261
O-Subject: [RHEL6.8 e-stor PATCH 57/58] hpsa: move lockup_detected attribute to host attr
Bugzilla: 1247729
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

This patch fixed the panic reported in BZ1259673 for RHEL7.2.
The fix is also required for RHEL6.8.

This patch was under review on linux-scsi mailing list when
it was added to this patch set.  It has since been added to
scsi.git tree and to kernel.org linux.git tree.
The following is from the description section of the patch
posted at:

http://www.spinics.net/lists/linux-scsi/msg88129.html

Original commit text is below:

This patch fixes a 'general protection fault' issue by
moving the attribute to where it was likely meant.

Signed-off-by: Tomas Henzl <thenzl@xxxxxxxxxx>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>